requirement.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
PyJWT==1.7.1
pymongo==3.10.1
six==1.14.0
Werkzeug==1.0.0
